sqlite versions of the filterBy operator
